Fixed massive bugs
