+++ Assets/Powers/Materials/UI/Tut/ui_tut_ctrl.mat
@@ -74,5 +74,5 @@ Material:
     - _UVSec: 0
     - _ZWrite: 0
     m_Colors:
-    - _Color: {r: 0, g: 0, b: 0, a: 0}
+    - _Color: {r: 0, g: 0, b: 0, a: 1.4e-44}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
