+++ Assets/Powers/Materials/UI/ui_loadingScreenFade.mat
@@ -74,5 +74,5 @@ Material:
     - _UVSec: 0
     - _ZWrite: 0
     m_Colors:
-    - _Color: {r: 0, g: 0, b: 0, a: 0.9994378}
+    - _Color: {r: 0, g: 0, b: 0, a: 0.9995068}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
